requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

six
WebOb
eventlet
oslo.db
oslo.log
oslo.i18n
oslo.utils
oslo.config
oslo.services
oslo.concurrency
oslo.serialization
greenlet
PasteDeploy
Routes!=2.0,>=1.12.3;python_version!='2.7'
netaddr